Manually tail-optimize the auto-restart
Summary:
Previously, we were using recursion to restart the
function. But we were heavily using dynamic bindings, error handlers
and restart cases, so nothing was being tail-call-optimized.

The easiest solution, it turns out, is to manually unwind the
tail-call-optimization. This requires the code to not be functional,
with a lot of registers needing to be set, but it does work.

The performance of all of this is questionable, but should be fine for
most use cases.

Differential Revision: https://phabricator.tdrhq.com/D8323

GitOrigin-RevId: 36a54834de552f92c1a3f3e13bd97166edd657bd